feat(amazonq): reword settings + import old CW settings (#4783)
* feat(amazonq): reword settings + import old CW settings

Problem: Some of the wording of the settings was off. Also, changing the name of the settings = a brand new setting in VS Code.

Solution: In order to preserve trust, we will try to import old settings users may have had. Extensions leave settings behind when they are uinstalled, so we can detect the old CW settings and import them into the new ones named after Amazon Q. This creates a seamless setting experience for migrating users.

* add .

* disable lint for long line

* update config change events

* use and update already existing migration function