fix(auth): delay PKCE server close #4780
Problem:
We assume the auth server page load-time will take less time than the
create token request. If that assumption is false, the auth server will
already be closed and won't be able to serve the UI.

Solution:
Delay for 5000ms.

Note: in practice this should actually never happen. This is just extra
insurance that if it does happen we are protected.