feat(amazon q transform): Human In The Loop #4607
Problem
Users builds fail sometimes do to 1p dependencies and those resources are not available at the transformation level in the backend. We are adding an "human in the loop" intervention to allow users to choose a version and resume a transformation

Solution
Display chat input fields, from maven output, to allow users to choose a 1p dependency upgrade version and continue on.