ui: remove jiggle from Login page items
There was previously a jiggle when moving the mouse on the
items since all border sizes were not the same.

Solution:

Make a single css class with the base item properties so that
regardless of state they all share the same border width.

Then the stateful classes like hover or selected will only change
what they need to change.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>