fix(codecatalyst): menu command fails after using tree node
Problem:
The codecatalyst commands:

    "Open Dev Environment"
    "Clone Repository"

fail if their menu forms are used AFTER the treenode forms.

Steps to reproduce:
1. click the "Open Dev Environment" tree node in the codecatalyst panel.
2. close the wizard.
3. in the codecatalyst panel, expand the "..." root menu and choose "Open CodeCatalyst Dev Environment"
4. it fails because the TreeNode from step (1) is passed as the first arg, which breaks `openDevEnv()`.

Solution:
Follow the instructions from the `VsCodeCommandArg` docstring:
https://github.com/aws/aws-toolkit-vscode/blob/4decdf6341a38c58db0e9f778ba57b5514434b0b/packages/core/src/shared/vscode/commands2.ts#L23-L41

TODO:
Note that the bug does NOT occur if you use the "..." menu command
FIRST. It seems like the TreeNode is stored as state on the
`CommandResource`; is that actually necessary?