fix(settings): only migrate telemtry setting if we are in amazon q (#4896)
We attempt to migrate in both extensions, even though we should only do so in the amazon q setting. The toolkit attempt fails, but it still pollutes the state with the stored keys.
This isn't expected to have an effect because the states of the 2 exts are presumed to be separate, but maybe there is a subtle bug revolving around stating here.