fix(amazonq): codeSelection incorrect if user has no selection #4921
Problem
Ask Q to write a unit test of a function, Q does not know how to do it. It hallucinates the function.( It does not know which class I am generating against). See internal TT P128119238.

Solution
Let the cursorState.range be undefined if user do not have a selection. The function getSelectionInsideExtendedCodeBlock is supposed to take the original selection as input, not the importantRange (which is current line of cursor)