fix: copyPackageJson script development mode
Problem:

With the recent addition for the null extension fix in debug mode,
we ran the hack during actual packaging of the AWS Toolkit extension.
This hack caused the vsix to not work

Solution:

Only run the hack during development. We did this by
adding a flag to the handlePackageJson script to do
the hack if the '--development' flag is given.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>