use new oidc client, remove our temp one
Since the new OIDC client with Authorization Code flow was
not released we had to use a pre-build version.

Now that the new client is released we can actually use it.

Solution:

This commit removes the old temp OIDC client and simply uses
our existing one with minor modifications to get things working.

Also we were required to update some transitive dependencies which
is why there are 'smithy' related dependency changes

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>