devEnv: try start activity heartbeat on connection change
Problem:

We only tried to start the activity heartbeat on extension startup. So if
that failed due to something like an invalid connection, the heartbeat mechanism
would not start

Solution:

Try and start the heartbeat mechanism when the CC connection changes as well.
But if it is already running then we just skip it.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>