fix(auth): use startUrl for cache key (#4983)
- Also adds startUrl to the registration file.

Problem: We only index client registrations files on region, scopes. IdC and builder ID are not distinguishable at the region + scopes level. This means that we can cache to the same registration file if the user tries IdC or builder Id. Now that we have moved to a new auth flow we no longer control opening the start URL, it is provided by the registration. If the registration is shared between IdC and builder ID, whatever one is accessed first will have its start URL saved to the cache.

Solution: Add start url to the hash key. This will prevent IdC and buidler ID logins matching to the same registration and routing to the same start url.

Steps to reproduce without this:
1. Try to log into builder ID or IdC
2. Don't fully sign in, cancel and return.
3. Try to sign in with the other.
4. The start URL will be for sign in method in step 1, not the one from step 3.