auth: do not use URI Handler in Auth PKCE flow (#4987)
Problem:

When users finish accepting scopes in the browser,
they are brought to a custom "success" page. Additionally
we use the URI Handler to then redirect them back to the IDE.

The issue is that the user is being redirected from the local server
address which includes a port number. Since this port number is
dynamic there will always be an "allow" prompt in the browser, instead
of seamlessly taking them back to the IDE.

Solution:

For now we will disable the URI Handler redirect and add it back in
if the vscode:// uri becomes allowed to use as a redirect URI instead

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>