dev: Edit Auth connections + fix stale auths
Problem:

It is common for want to view/modify the Auth Profile store.
The Profile Store is our source of truth for auth connections that
the extension is aware of and using.
We sometimes want to read and modify this source of truth for debugging
purposes.

The issue is that this option is hard to find since you need to dig through
multiple menus. Also once you get to it, the content you see will become stale
if it every changes.

Solution:

- Add a new item to our Developer Menu to open the auth profile store editor.
- Fix issue where the text document editor was not refreshing with the latest changes.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>