codecatalyst: fix invalid SSO profile in Dev Env
Problem:

When a Dev Env is spun up, the AWS Toolkit extension automatically creates
an SSO profile in the Dev Env using the the sso session metadata in the ~/.aws/config file.

In the scenario the ~/.aws/config file contained the session for a non-builder id session
the SSO profile created would be incorrect due to the startUrl and Region being reversed
with eachother when the SSO profile was created.

This led to an invalid SSO profile, so any attempts to reauthenticate would fail.

Solution:

correctly create the SSO profile.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>