fix(auth): enforce 5 scopes for using existing connections in amazon q
Problem: Users with 3 scopes can import their connection in amazon q (it is not autoconnected, it must be selected manually in the login ui). This creates edge cases and creates non-unformity for connection users. Q features may behave in unexpected ways.

Solution: Instead of checking for just 1 Q scope, check for all 5. Importing a connection was intended to add missing scopes anyways.