auth: use minimal CW scopes with Q, but expire connection
Problem:

In AWS Toolkit 2.19 we only added the 3 CW scopes for BiD, but now with Amazon Q there
are an additional 2 scopes (gumby/weaverbird).

The issue is that connections made with 2.19 did not seamlessly transition in to the new
standalone Q extension. Instead users needed to know to go to the Q login webview and select
their existing BiD connection and have it reauthed to work with Q.

Instead of the above behavior, if the user has the 3 scope connection, we want to auto connect
to this connection but then mark it as expired so that the status bar icon for Q tells the
user they need to reauth (yellow background and reauth icon).

Repro Steps of current state (before this change):
- Delete all connection cache from AWS + Amazon Q extensions first
- Download AWS Toolkit 2.19 (this only adds the 3 scopes)
- Connect to CW with BiD
- Install the latest Amazon Q

Note how the old connection is not automatically transitioned over, the status bar shows the 'X' as there is
no connection. But if you open the Q login webview it will show BiD, then you can select that and things will
work again.

Solution:

This commit changes the code to detect a 3 scope CW connection, then use it for Q but mark
it as expired.

So what will happen is that if you follow the repro steps above, but instead launch Amazon Q
with this change, once you open the Q webview it will auto connect to the old connection but
show it as expired in the status bar.

Things that still need to be addressed:

- The logic to check + use the 3 scope connection only triggers when the Amazon Q panel/webview
  is opened. Until it is opened nothing happens, and the user is not made aware to reauth in the
  status bar icon. A better solution is to move this logic to the backend code so we can trigger it
  without the need to open the webview. Or does it make sense to auto open the Q webview if we detect
  they only have the 3 scopes instead of 5

- If the status bar is shown as expired, then none of Q chat will work BUT the user will still be
  able to type messages in that (please verify this). Though it will say they need to reauth once they send the message.
  Do we want this behavior or to show the reauth page?

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>
Co-authored-by: Maxim Hayes <149123719+hayemaxi@users.noreply.github.com>