refactor: make timeout work in web
Problem:

`globals.clock.setTimeout()` creates a timeout but returns a different
result depending on Node.JS vs Web. In Node.JS it returns an object, but
in Web it returns a number.

Some places call `refresh()` on the result, this works in Node.JS but fails
in browser.

Solution:

Update the Timeout class so that when `refresh()` is called it figures out
which environment we are in and does the appropriate refresh.

Also, update functionUtils to use the `Timeout` class instead of directly using
globals.clock.setTimeout()

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>