lint: catch incorrect usages of once(), fix current usages
Problem: `once()` can be used incorrectly such that code can be executed multiple times, even if it seems that that the developer just wanted it to run a single time.

Solution: Custom lint rule to try to catch these cases. It is not foolproof, but it should catch most issues.