fix(csharpSamDebug): remove debug mode for arm64 #4934
Problem:
When `sam local` is invoked for `cSharp`, the invocation gets blocked when
trying to attach a debugger. `vsdbg` is not supported for `arm64` architecture. 

Solution:
Disable `debug` mode for cases when the template architecture is `arm64` or the
developer is using `arm64` architecture machines, for eg - Apple M1/M2/M3 chips.