build: remove --vueHr flag for amazonq debugging
Problem: arg should go to npm run copyFiles in buildScripts, but it isn't and hasn't been. Now with the addition it of tsc -p ./ --noEmit it is going to that instead and it's an unrecognized arg.

Solution: remove it

Not sure if the arg ever worked. We can revisit this if the feature is desirable.