telemetry(Amazon Q Code Transform): move metric earlier in transformation flow #5032
Problem
`jobStartedCompleteFromPopupDialog` telemetry was supposed to signify one of the first events happening in the transform flow:
```
A validated job was started but we have not yet compiled using Maven.
```
However currently its emitted after `mvn clean install`, this is wrong. 

Solution
Move the `jobStartedCompleteFromPopupDialog` telemetry after the `prevalidation` is finished but before the `mvn clean install`.