fix(auth): login page displays connections that already exist for that service (#5186)
Problem: We do not allow the SSO form to continue if the user tries putting in an IdC start url that already exists in the same extension. However, this check isn't in place for displaying connections from the other extension. This means users can duplicate the connection start url in the extension and caused undefined behavior.

Solution: Filter out existing connections from other extensions on what connections are in the current extension.