auth: ignore vscode proxy errors on token refresh #5198
We were seeing an error in `aws_refreshCredentials` metric.
This error was networking related when the user has setup a proxy
in vscode, but they were not connected to it. Maybe they didn't
connect to VPN.

We typically catch network errors during the SSO token refresh process
so that the session does not become invalidated on it. We do not catch
all errors, but will add them in incrementally as we see them in telemetry.
This proxy error was not caught and caused sessions to become invalidated prematurely.

Solution

Treat this VSCode Proxy Error as a network error so that we ignore
it during token refresh

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>