auth: add telemetry for why the user needed to reauth
Problem:

In the reauth metric (aws_loginWithBrowser + `isReAuth: true`) we didn't know why
the user needed to reauth. It could have been due to regular expiration or some unexpected
error.

Solution:

Add a new field in `aws_loginWithBrowser` named `reAuthReason` which will have some sort
of identifier for why reauth was needed.

On a technical level, we save the reason an SSO connection was invalidated in our cache
using the ReAuthReasonState class. Then when it comes time to do a reauth we will grab
the past value from that state and set it in the metric.

The reauth reason can be set in a few places, but some examples are when the token refresh
process fails, we will know that this is the cause for the user needing to reauth.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>