web: make 'common' codewhisperer export module
Problem:

When importing the codewhisperer exports module from index.ts in web mode
it was breaking due to a transitive fs-extra import.

Solution:

Export a separate module from indexCommon.ts for code that works in web
or node. This file is a subset of index.ts, so behavior remains the same.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>