refactor: start moving services to `awsService`
Problem:

We have multiple aws services that each have their own folder in the root.
This adds more folder to have to search through in the top level

Solution:

Create an `awsService` folder and start the move of the existing folders
in to them. This commit moves CDK for now, and in future commits we will move
the rest.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>