fix(appcomposer): add warning message to version 1.91.0 about drag and drop
Problem:
VS Code version 1.91.0 has a bug causing the dragEnd action to not trigger in WebViews unless Shift is held down.

Solution:
This has already been reverted and should be fixed in the next VS Code version. In the meantime, this adds a warning message informing users of the workaround.