Add postMessage embedded APIs; set isEmbedded global
Problem: sendMessage and receiveMessage VSCode
proposed APIs, which are made available on VSCode
for Web via a fork of the core code-oss project,
need to be accessible to the AWS Toolkit.
Additionally, Toolkit should be aware of when it
is being run in an embedded context.

Solution: detect if sendMessage and receiveMessage
APIs are available on extension activate. If they
are available, VSCode is running in a context
which supports these two APIs.

In a future commit, we will add support for a
unique auth strategy when embedded mode is
detected.