auth: createToken accepts args related to reauthentication
Problem:

SsoAccessTokenProvider.createToken() is used for both authentication and reauthentication, but
there isn't currently any way to identify the reason. This leaves a gap
in our telemetry.

Solution:

Created an object CreateTokenArgs which will hold args to be used by
createToken(). Now we can pass data to other parts of SsoAccessTokenProvider
that may want this data.

Most of the changes are adding in this new argument to createToken(). The notable
change is in auth.ts#Auth.reauthenticat() where we now pass isReAuth in the createToken() call.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>