refactor: remove hacky telemetry for reauth
Now that we've changed how we pass the isReAuth and reAuthReason
values to the final telemetry metric, we can get rid of the hack
we used to pass these values in Auth.reauthenticate().

This commit is removing redundant code. SsoAccessTokenProvider.withBrowserLoginTelemetry()
essentially has all the information it needs and doesn't need Auth.reauthenticate()
to use `record()` to externally setup values.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>