isEmbedded should also rely on isWeb
There is no case where isEmbedded should be true when isWeb is false, since the sendMessage and receiveMessage APIs are a wrapper around the browser postMessage API. There will however, be cases where both are false, or only isWeb is true. This change means we eliminate the impossible state of isWeb = false and isEmbedded = true.

Co-authored-by: Nikolas Komonen <118216176+nkomonen-amazon@users.noreply.github.com>