fix(amazonq): always show latest build logs #5316
Problem:
When users download the pre-build error logs, then run another transformation
without clicking the "Start a new transformation" button, if another pre-build
error occurs, the **old** logs will show up. This could happen if they close the
Q Chat tab and re-run a transformation, or if they close the IDE between job
runs. If they do click the "Start a new transformation" button, everything works
fine, but we need to handle these other situations.

Solution:
Use `makeTemporaryToolkitFolder()` which creates a new folder with a random
name, so that the logs are downloaded there, meaning we don't have to deal with
deleting the old logs that would otherwise be in the same folder as the new
logs.