fix(amazonq): fix index cache not loaded when IDE restarts
Fix index cache not loaded when IDE restarts.

The boolean field `refresh` of Index request tells LSP whether to rebuild the index completely (force refresh). When it is set to false, the LSP always loads cached index (if it is not more than 2 days old). Refresh should be set to true only if the IDE has absolute reason to force the LSP re-calculate the index.