auth: add telemetry for why the user needed to reauth (#5243)
* util: Add in-memory cache utils interface

Problem:

There is a common use for giving a key and getting back a cached
value. We have an existing implementation named KeyedCache, but it
utilizes Promises, leaving it open to race conditions.

Solution:

Create a synchronous interface similar to KeyedCache.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* auth: createToken accepts args related to reauthentication

Problem:

SsoAccessTokenProvider.createToken() is used for both authentication and reauthentication, but
there isn't currently any way to identify the reason. This leaves a gap
in our telemetry.

Solution:

Created an object CreateTokenArgs which will hold args to be used by
createToken(). Now we can pass data to other parts of SsoAccessTokenProvider
that may want this data.

Most of the changes are adding in this new argument to createToken(). The notable
change is in auth.ts#Auth.reauthenticat() where we now pass isReAuth in the createToken() call.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* auth: add telemetry for why the user needed to reauth

Problem:

In the reauth metric (aws_loginWithBrowser + `isReAuth: true`) we didn't know why
the user needed to reauth. It could have been due to regular expiration or some unexpected
error.

Solution:

- When we invalidate an SSO connection, record the reason
  in ReAuthState. Then next time we reauth we can look at this state to see why.
- The ReAuthState is in the SsoAccessToken provider since invalidation happens
  in this class itself and not in Auth. The SsoAccessTokenProvider.invalidate()
  is called to invalidate a connection and knows why the connection was invalidated.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* refactor: remove hacky telemetry for reauth

Now that we've changed how we pass the isReAuth and reAuthReason
values to the final telemetry metric, we can get rid of the hack
we used to pass these values in Auth.reauthenticate().

This commit is removing redundant code. SsoAccessTokenProvider.withBrowserLoginTelemetry()
essentially has all the information it needs and doesn't need Auth.reauthenticate()
to use `record()` to externally setup values.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* run formatter

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* refactor: fix PR comments

- Add some comments regarding why we call 'MapSync' with 'sync' in the name
- Rename 'MemoryMap' to 'NestedMap'
- rename some of the Map interface functions to be closer to the ES Map interface

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* refactor: merge new reauth state/telemetry tests with existing tests

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

* rename: asKey() -> hash()

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>

---------

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>