telemetry: include chained error message in reasonDesc
Problem:

When we use the function getTelemetryReasonDesc() we would lose information
about the underlying error in the case the error was a ToolkitError
with an underlying chained error. Only the message of the ToolkitError
was used.

So in telemetry we would be missing some information about the real cause
of the error.

Solution:

getTelemetryReasonDesc() will return a  message which consists
of all messages in the chain. It recursively calls in to the next
chained error of a ToolkitError and appends it to the final output.

This will give us some context about the higher level call as well
as the root of the issue.

The final message could look something like:
"Message A::Message B::Message C"

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>