refactor: make an atomic write (#5361)
Problem:

When writing to our SSO cache we noticed malformed json files that had
our token. As a solution we did an "atomic" write where we wrote a temp
file then renamed. This prevented the content from being malformed due to
a separate write.

The issue is that the combination of writing a temp file, then renaming it
fails sometimes as indicated by our metric `aws_refreshCredentials` where we
were seeing a 'FileNotFound' error.

Solution:

Move the atomic write logic in to `FileSystem.writeFile()` itself, where it is
triggered as part of an argument to the function.

Additionally when the `rename()` method is called under the hood of `writeFile()`,
we validate that the file exists by retrying a few times. We suspect this to be
the culprit of the `FileNotFound` error, see the comments in code for more info.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>