bug: Amazon Q sign in blocked by stale connection
Problem:

Previously the AWS Toolkit and Amazon Q extensions shared Auth connections.
To do this they communicated their Auth connections to each other, with each
extension storing this in their auth connection cache.

Later we decided to stop sharing the Auth connections between the extensions,
instead having each extension create its own connection.

The problem is that when we created this separate auth connection change, we
had stale connections remaining in Amazon Q. This caused issues like users being
unable to sign in to a new connection since the cached connections showed
it already existed.

Solution:

On every startup clear the stale connections.
We have existing code that does this, but it only ran in a CodeCatalyst dev environment.
But it looks like we always want to run this check as it seems there are other cases
where this happens.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>