refactor: increase scrubNames() file ext limit
Problem:

scrubNames() removes PII from a file path. There is also a limit
to the file extension length of 4 characters. If the length is greater,
the file ext is fully excluded.

We have some scenarios in telemetry where the file has no file ext
but we want to see it. A guess is that there are some cases where the
file ext is longer than 4 chars.

Solution:

Show the file ext regardless of its length.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>