bug: rename() fails causing FileNotFoundError
Problem:

In our aws_refreshCredentials telemetry we saw a spike in FileNotFound
errors when writing the refreshed token to our filesystem.
This is caused by something in the vscode.workspace.fs.rename()
implementation that is not obvious and does not seem like an obvious error
on our end. Most of the time things work, so most users are not running
in to this problem.

Originally, to fix some errors happening during writing to the filesystem
we created an "atomic" write, but there are edge cases where this did
not work. So there are different implementations of writing a file that we used
to rememdy this.

Solution:

Try the next fallback methods for saving the file to the file system when one fails.
See the comment in the code for the explanation of the implementation.

We also report telemetry on the specific failed cases to hopefully get a better understanding
of the specific failures. We can use this

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>