fix(telemetry): try to handle SyntaxError for createToken requests
Problem: Rare and unreproducible case where SDK client calls to create an SSO token return an html page instead of JSON data. It is probably some error page indicating an interference with the request.
Solution: Try to extract the reason and status code from the page if possible, and report it in telemetry.