internal data collection for chat workspace (#5330)
## Problem

The relevancy of workspace context retrieved from the LSP has unknown quality because of lack of data.

## Solution

As required, by default enable workspace index for a fraction of amzn users and always send relevantDocument to the service. There is no customer facing changes. This change is purely for data collection and will be removed soon.

To only apply this to a fraction of amzn users, we rely on the listFeatureEvaluations API in the backend.

Tested with builder id login, amzn login, with and without @workspace.