ignore venv when local indexing (#5433)
## Problem

When triggering the indexing of a project, if the .venv or venv directories from a python virtual environment is not listed in the .gitignore file, the workspace feature tries to index all the files in those directories causing it to take a really long time OR fail due to size.

## Solution

Ignore .venv and venv at local indexing. We will build more robust ignore mechanism in the LSP soon. This is a stopgap for now.