fix(amazonq): avoid refresh suggestion for paginated response (#5437)
## Problem

A flash exists in codewhisperer inline because after each pagination call we refresh the max number of suggestions in the tooltip.(1/1 to 1/5, then 1/6) Since we refresh the max count, we also refresh the suggestion itself. VSC do not have API to let us just refresh the max count.(we fresh it by re-creating the inline completion object)

## Solution

Remove the refresh of index. This solution has been approved by product.