docs: minimize PULL_REQUEST_TEMPLATE.md
Problem:
For better or worse, PRs tend to be squash-merged, so devs tend to
neglect commit messages, and the PR description (as opposed to commit
messages) contains the relevant info.

Therefore to encourage a useful git commit history, we have configured
the github repo to use the PR description as the "Default commit
message".

But the `PULL_REQUEST_TEMPLATE.md` contains a lot of boilerplate that
would then get pasted into the merged commit message unless the merging
party takes time to delete it.

Solution:
Minimize `PULL_REQUEST_TEMPLATE.md` boilerplate to avoid lots of noise
in the git commit messages.