fix(amazonq): allow symlinks for JDK path (#5500)
## Problem

Some users provide a symlink when QCT prompts them for a JDK path, and
this previously was not being accepted.


## Solution

Remove part of a conditional check (`isDirectory`, since that is always
false for symlinks). This allows users to use symlinks, which will work
fine during the local build. Even if someone provides a path that is
*not* a directory now, the local build will fail and build error logs
will open up for them to debug, as usual.

---

<!--- REMINDER: Ensure that your PR meets the guidelines in
CONTRIBUTING.md -->

License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

Co-authored-by: David Hasani <davhasan@amazon.com>