telemetry: Add `sessionId` to metrics
Problem:

I have filtered all the metrics in Kibana by `clientId: "XXXXXX"`, but if they had multiple IDE windows opened at the same time sending metrics I cannot know which window they are coming from since they share the same clientID.

Solution:

This adds a new field to the telemetry service  called `sessionId` which uniquely identifies
a part of the UI in an IDE window. It is adjacent to `clientId`.

Now in my metrics I will be able to differentiate metrics by their sessionID
since they both share the same clientId by design.

Implementation Details:

Changes in the Telemetry service need to be merged before this change can since they update
the endpoint to accept this new field.

The updates of this commit only mirror the upstream, so this process is manual.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>