fix(auth): SyntaxError not being considered Network Error
Problem:

There were some previous changes that caused the existing implementation
of a SyntaxError to not be considered a Network Error. And during token
refresh we were seeing SyntaxErrors still appearing

Solution:

Create a single class which wraps a `SyntaxError`, called `SDKClientResponseError`.
Under the hood of a SyntaxError is an error that originates from a failed AWS SDK Client response.

This new class can only be created if given the correct SyntaxError instance
to `SDKCLientResponseError.fromErrorIf()`.

Any existing code that was related to this scenario has all been centralized in to
the SDKClientResponseError class.

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>