fine tune criteria of a SyntaxError that is an AwsClientResponseError
We were previously seeing if a $response.reason field existed in
a SyntaxError to determine if it was an AwsClientResponseError.
But this was prone to potential false positives.

So this commit makes it even more specific requiring that the
$response.reason message itself met certain criteria. And based
on the telemetry we saw it typically started with:
"SDK Client unexpected error response: data response code: ..."

Signed-off-by: Nikolas Komonen <nkomonen@amazon.com>